sideproject/sideproject6_multitimer/multitimer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
qrcode==7.4.2
Pillow==10.1.0
redis==5.0.1
python-dotenv==1.0.0 